LIBIIIF-151. Use file locks to prevent concurrency issues.
When reading and writing the mezzanine files, multiple requests may come in before the first request is able to retrieve the origin image and complete the conversion. Adding a file lock to each mezzanine file means that only one thread may access that file at a time. If there is no local file, the first thread does the retrieval and conversion. Subsequent threads requesting the same file block until that is complete then each can access the file one by one for reading.

https://issues.umd.edu/browse/LIBIIIF-151